sbir-grants/mcp-server/pyproject.toml
[project]
name = "sbir-data-server"
version = "0.1.0"
description = "MCP Server for SBIR market data aggregation"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "httpx>=0.25.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "python-docx>=0.8.11",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.2"
]

[project.scripts]
sbir-data-server = "server:main"

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
